requirements.txt
gdown==5.2.0
matplotlib==3.7.1
mmengine==0.10.7
mmseg==1.3.0
numpy==1.24.3
opencv_python==4.10.0.84
pandas==1.5.3
Pillow==9.4.0
Pillow==12.0.0
Requests==2.32.5
scipy==1.10.1
timm==0.9.10
torch==2.1.0+cu121
torchmetrics==1.2.1
torchvision==0.16.0+cu121
tqdm==4.65.0
